ingestion/requirements.txt
# BITS-SIEM Ingestion Service Requirements
# Core dependencies for multi-protocol syslog ingestion

# Async framework
asyncio==3.4.3
uvloop==0.19.0

# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1

# Message parsing and validation
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.0

# Geo-location services
geoip2==4.8.0
requests==2.31.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration management
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
